requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
plotly
streamlit
joblib
openpyxl
pymongo
dnspython
python-dotenv